refactor(#449): align PGID probe + codex failure formatting with #465 patterns
